refactor: Rename lib.twitter to lib.twitter_api
