+++ app/utils/insert/lookupAndStoreTweets.py
@@ -17,7 +17,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(
     os.path.dirname(__file__), os.path.pardir, os.path.pardir)
 ))
 from lib import tweets
-from lib.twitter import auth
+from lib.twitter_api import auth
 
 
 def main():
